Added base part of AI
I created classes, which i will be use in future for implement full AI logic.
I create a navmeshBoundsVolume on level. After plane and bounds initialization - we scale this volume and update it. In my opinion, this is a questionable decision, but in the current situation, it is quick and effective.